# docs/contracts/api.openapi.yaml
openapi: 3.1.0
info:
  title: VideoCutlist API
  version: 1.0.0
  description: >-
    Deployment authentication applies to application endpoints. API responses carry
    a server-generated X-Request-ID; errors contain error.code, error.message, and
    matching error.requestId. Missing/invalid authentication is 401, rejected origins
    are 403 origin_forbidden, and unexpected persistence/runtime failures are safe
    500 responses. Existing 422 input-validation exceptions are retained. HEAD
    responses have no body. MCP JSON-RPC is a separate protocol outside this contract.
servers:
  - url: "{serverBaseUrl}/api/v1"
    variables:
      serverBaseUrl:
        default: http://localhost:8787
        description: >-
          Reachable VideoCutlist server base URL; transport provider
          independent.
security:
  - bearerAuth: []
  - {}
paths:
...
  /settings:
    get:
      summary: Read runtime settings and safe root diagnostics
      operationId: getSettings
      responses:
        "200":
          description: >-
            Effective settings with safe labels; raw paths are never returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsResponse"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
    put:
      summary: Replace runtime settings
      operationId: updateSettings
      description: >-
        Complete mutable settings document, at most 1048576 bytes. Deployment roots
        and destinations remain server-owned and may be omitted; destination metadata
        may be echoed unchanged. Invalid settings (including exportLimit outside
        1..64) are rejected with 422 before live mutation or persistence. Stale
        revisions are 409. Unexpected application/persistence/rollback failures
        are 500. Live settings affect new work; existing export snapshots retain
        their settings, except shared worker/process/cache limits. Preview defaults,
        scan limits, and MCP enablement apply live. MCP transport request capacity
        remains the startup value; previewGlobalLimit controls media processes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SettingsUpdated"}
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Used only when the server selects static bearer mode.
...
  responses:
    Error:
      description: Safe error; the request ID matches X-Request-ID
      headers:
        X-Request-ID: {schema: {type: string}}
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    SettingsUpdated:
      type: object
      required: [settings, revision, schemaVersion, updatedAt]
      properties:
        settings: {$ref: "#/components/schemas/RuntimeSettings"}
        revision: {type: integer, minimum: 1}
        schemaVersion: {type: integer}
        updatedAt: {type: string, format: date-time}
    SettingsUpdate:
      type: object
      additionalProperties: false
      required: [revision, settings]
      properties:
        revision:
          type: integer
          minimum: 1
        settings:
          allOf:
            - $ref: "#/components/schemas/RuntimeSettings"
            - type: object
              required: [exportLimit, cacheMaxBytes, previewGlobalLimit, previewBeforeMs, previewAfterMs, previewMaxMs, previewGridMs, mediaMaxFiles, mediaMaxDepth, mcpEnabled]
              # Explicit types keep required PUT fields non-optional in generated clients.
              properties:
                exportLimit: {type: integer}
                cacheMaxBytes: {type: integer}
                previewGlobalLimit: {type: integer}
                previewBeforeMs: {type: integer}
                previewAfterMs: {type: integer}
                previewMaxMs: {type: integer}
                previewGridMs: {type: integer}
                mediaMaxFiles: {type: integer}
                mediaMaxDepth: {type: integer}
                mcpEnabled: {type: boolean}
    RuntimeDestination:
      type: object
      required: [id, label, kind]
      properties:
        id: {type: string}
        label: {type: string}
        description: {type: string}
        kind: {type: string, enum: [download, archive, source_adjacent]}
        retention: {type: string}
    RuntimeSettings:
      type: object
      additionalProperties: false
      required: [exportLimit, cacheMaxBytes, previewGlobalLimit, previewBeforeMs, previewAfterMs, previewMaxMs, previewGridMs, mediaMaxFiles, mediaMaxDepth, mcpEnabled]
      properties:
        destinations:
          type: array
          items:
            $ref: "#/components/schemas/RuntimeDestination"
        exportLimit:
          type: integer
          minimum: 1
          maximum: 64
          description: >-
            Live shared scheduler worker limit for export, detection, and library
            scan jobs. 64 is a safety ceiling, not recommended concurrency.
            Admission allows at most four times this limit in queued/running jobs.
            Increases allow more workers immediately. Decreases do not cancel
            active jobs or discard queued work; excess workers retire after their
            current job and queued work drains at the new limit.
        cacheMaxBytes:
          type: integer
          minimum: 1
          description: >-
            Live preview-cache total budget and timeline-asset per-artifact limit
            (not a combined cache budget). Subsequent publications use the current
            limit, including in-flight generations. Oversized timeline cache entries
            are misses. Existing preview entries are evicted on subsequent commits.
        previewGlobalLimit:
          type: integer
        previewBeforeMs:
          type: integer
        previewAfterMs:
          type: integer
        previewMaxMs:
          type: integer
        previewGridMs:
          type: integer
        mediaMaxFiles:
          type: integer
        mediaMaxDepth:
          type: integer
        mcpEnabled:
          type: boolean
...
    SettingsResponse:
      type: object
      required:
        - settings
        - revision
        - schemaVersion
        - updatedAt
        - pathsConstrained
        - roots
      properties:
        settings:
          $ref: "#/components/schemas/RuntimeSettings"
        revision:
          type: integer
        schemaVersion:
          type: integer
        updatedAt:
          type: string
          format: date-time
        pathsConstrained:
          type: boolean
        roots:
          type: object
          additionalProperties:
            type: object
            properties:
              state:
                type: string
                enum: [ready, unavailable]
              message:
                type: string
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - requestId
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
...